+++ Matrix.h
@@ -22,7 +22,7 @@ public:
 	void checkIfEqual(Matrix &other) const;
 	void add(Matrix other);
 	void print() const;
-	void setIndexValue(int i, int j, int value);
+	void setIndexValue(int i, int j, double value);
 	Matrix filter_slide(Matrix filter, int stride, int bias);
 private:
 	int height;
